Feature: allow dynamic filter swaps (#4)
This PR adds dynamic filter swaps by storing the Filters in an Arc<_> which then is stored in an ArcSwap.

Arc prevents invalidation of in-use memory by managing the lifetime of the object. It is generally not important that threads are aware of the new filter as soon as it updates, just that they don't try to access deallocated memory.

ArcSwap is used for interior mutability, it allows us to swap the internal Arc with a new value in an atomic manner and without locks, while ensuring any threads that might have accessed the data still have access to valid memory by virtue of relying on Arcs.


Co-authored-by: Jonathan Schwender <55576758+jschwe@users.noreply.github.com>